Make load-compilation-database safer, closes #92 via monkey-patching that particular case
i.e. clear *db* if the db contains are directories which do not exist.

Don't cherry-pick this without the preceding commit which adds (assert *db*) to LOAD-ASTS; otherwise, it's a catch-22, damned if you "continue unsafely" (due to the fatal exception from llvm), damned if you don't (due to segmentation fault).