Moved gc_interface.* files to src/gctools
I moved the gc_interface.* files to src/gctools because
I'm trying to reduce the complexity of the src/main directory.
The contents of the src/main directory need to be duplicated for
different extended versions of clasp such as Cando.  The less stuff in
the src/main directory, the less duplicate code there is.