quicklisp works fixing #156
stassats found that (funcall (compile nil '(lambda (x) (- x))) 1) didn't
work. I tracked it down to a bad define-compiler-macro in inline.lisp
that I fixed in one of the preceeding commits.  In this commit I fixed a
bug when a simple-stream-error was signaled and I added support for
the name-char "null".