Several changes
Upgraded to llvm-ToT October 21, 2016
Fixed calls to intrinsics so that they use 'call' rather than 'invoke'
because 'invoke' can only be used with a landing pad and landing pads
are created in a different way in cclasp than they are in bclasp.
Also changed some intrinsic functions from primitive to
primitive-nounwind.