Restored from_object/to_object in array.h
I think there is a reason to have dedicated from_object/to_object
translators in the array classes.  They are designed to as quickly as
possible, with no bounds checking, whatever representation is in the
array to and from a T_sp object.

I tried switching to the translate from_object/to_object and that
lead to problems that would take time to sort out.   I think it was
template specialization resolution issues.