Deal with inlining functions in non-null lexical environments
(pertains only to global functions defined by defun)

Inlining works by compiling the source form down to an AST. Previously
this compiling was done in the global environment, so anything outside
of the defun was ignored. Because unknown variables and functions are
silently ignored (considered special and assumed to be defined later
respectively) this would result in pretty wrong behavior.

Now, instead, the compile-time part of the environment is extracted and
used. So (macrolet ... (defun ...)) works as expected and results in an
inlinable function. (let ((x ...)) (defun ...)) does not result in an
inlinable function, i.e. the AST is not saved, so future calls are just
calls and that works.

This relies on cleavir-env:compile-time, which I wrote just for this and
is nowhere right now, so do not incorporate these changes until clasp's
cleavir has that function.