Use cleavir's new dynamic-extent understanding
Sadly this doesn't actually improve clasp's dynamic-extent. To do that
the mark-dynamic-extent function needs to be called, as is commented out
in translate.lisp. Using it works fine in cclasp if loaded in
separately, but if used for build clasp will segfault on boot for
unclear reasons. Working on it.

Also, at this point there are two distinct liveness systems, both of
which are used. Type inference should be updated; again, working on it.