Cull incorrect uses of coerce-to-function
Almost every library function that took a function designator used
coerce-to-function. c-t-f implements the behavior of (coerce x 'function),
including looking up setf function names and evaluating lambda expressions,
neither of which are included in the proper behavior of function designators.
So this is a bit of a standards conformance cleanup.

The sole use of c-t-f is now in coerce. Because of this I would prefer that the
C++ function be eliminated and replaced with a lisp version. The only reason I
didn't is that I'm not confident in my understanding of how to make an
interpreted closure from Lisp.