Fix build: vector-push-extend does not call adjust-array
Build was broken because something called vector-push-extend, which called
adjust-array if it was defined. During build it's possible for adjust-array to
be defined but for functions it calls not to be defined, etc., so there were
errors.

This fix isn't good, though. I don't think the "internal" function used now
fully replicates the behavior of adjust-array.

Beats build crashing though.