Have compute-effective-method return forms instead of functions
This does not work with clasp's build system. bclasp can be built, but not
cclasp.

Before this commit, the standard method combination was done specially, using
function standard-compute-effective-method. This function never called the
compiler, instead returning closures, in the general case one made by
standard-main-effective-method. So most calls to compute-effective-method never
hit the compiler.

This commit unifies things so that compute-effective-method returns a form
(which was the entire point - so that we could look at it and do optimizations)
which then has to be converted into an actual function somehow. For that we use
the compiler, for obvious reasons.

Now, in the course of compiling cclasp we build cleavir, which uses CLOS
heavily. Then we switch to it as our compiler and build it again. Because we do
all this in the same image/environment, this means that the second build is also
a redefinition.

The build gets to compiling general-purpose-asts.lisp and hits its first
DEFCLASS for a cleavir class. It has to redefine the class AST, and removes its
accessor methods. Removed methods trigger invalidation in the accessors, which
results in compute-effective-method calls. These trigger the compiler. Because
the class or accessors are being redefined, they are not available. The
compiler signals an unbound function error or suchlike.

I can't see any way to reconcile with the redefinition-based system. If we don't
use the compiler to compile this part of the compiler, the compiler will be
slower. The redefinition must be done because DEFCLASS has compile time effects.
Bla bla bla

I'm going to poke away at the thing needed for major build system revision and
hating