Moved the additional_data to follow header
MPS locks up if DEBUG_GUARD is on and
additional_data[0] isn't right after the header.
This happens because pad objects are at least two words and the
size is written into additional_data[0] and if its not right after the
header then it will intrude into another memory page and lock up.