Fix initarg cache invalidation
When a method on initialize-instance or whatever is defined, it should
invalidate the caches of (at least) the class the method is specialized on.
The previous behavior was to invalidate it for subclasses of CLASS, which is
totally wrong, as most classes are rather instances of subclasses of class.

Demonstration of the bug fixed:

(defclass foo () ())
(make-instance 'foo) ; so the cache is put in place
(defmethod initialize-instance :after ((o foo) &key test) (print test))
(make-instance 'foo :test "hello") ; -> an error, which is wrong.