Fix float to integer conversions (Fixes #399)
Numeric limits of int is just wrong, so forget that.

Just changing it to fixnum doesn't quite work either. As it says in the
comment, most-negative-fixnum is a negated power of two, while
most-positive-fixnum is slightly less than a power of two. For a floating point
format based in binary (like IEEE 754's), this means most-negative-fixnum can
be represented exactly, and most-positive-fixnum cannot be. If (double)mpf is
cast back into an integer it will be higher than mpf. The next float down is
lower than mpf by some, so we use < for the high end.

Since most-negative-fixnum can be represented exactly,
(fixnum)(double)mnf = mnf
no problem. So <= for the low end.

In clasp we have to make sure that there are no bignums within the fixnum range.
of constants from md5, being stuffed into an ub32 array. The values returned by
truncate would sometimes be 32 bit integers that were nonetheless bignums.
Our use of aset does not check whether its 32 bit integers are fixnums (because
this is supposed to be true by definition) so it wrote in addresses as if they
were numbers, resulting in a very inconsistent bug.