Rewrite dtrees to merge ranges properly
singles are removed in favor of degenerate ranges.

The previous code doesn't seem to have ever worked. This wasn't an issue
because outcomes were always distinct, and therefore no mergers were done.

This code MAY have fixed the problem. The problem did recur when the binding
in optimize-etc was (working (first class-specializers)) instead of
(working (first sorted)). This implies that the class specializers don't
come in sorted.

Ah and for the record the problem was a message in the fastgf debug about
keys already being in the call history - basically a false negative hitting
the slow case.