Enable fast methods
Fast methods means some methods having special "fast method functions" (FMFs).
These FMFs take the same arguments the method does, rather than like the usual
method functions which take a (vas)list of arguments and a list of next
methods.

Implementationally, this means the standard method on make-method-lambda
returns special initializers, and then fastgf checks for an FMF to use.
The former has several problems that are explained in comments. The latter is
mostly fine, except that the codegen is a bit hairy, see comment in cmpfastgf.