Avoid invalid caching related to eql specializers in CLOS
meister hit a bug in cst-to-ast. This commit may be redundant.
Basically, if a gf has a method with a class specializer and another
with an eql specializer, and you call it so that it hits the first method,
and the eql specializer object is a member of the class, the discriminator
is set up improperly. I have a fix in mind but first let's go back to
correct behavior.