Fix inlined aref
Before this I was hanging the system when doing basic things,
such as (jit-function-name '(method foo)). No idea why.

There was a small bug in inline where it would loop infinitely
due to using the wrong variable name. That didn't seem to be the
cause, but I'm not sure what else it could be. I was getting hangs
even with that fixed, so I don't know why this works.