MAKE-INSTANCES-OBSOLETE fix
there was a stray method and we invalidated at the wrong time

drmeister hit a weird problem where the class-slots in the stray
method reported the SLOTS slot was unbound. Have not reproduced,
don't understand how it would occur, but hopefully it won't recur,
if only because we no longer need slots there anyway.