Avoid ENSURE-GENERIC-FUNCTION when macroexpanding DEFMETHOD
See #487. Doing so has unintuitive consequences, so while allowed I
think removing it is fine. Especially since hardly anyone uses
make-method-lambda anyway, let alone in such a way as to check the
name of the generic function or suchlike.

Code is thanks to jackdaniel, who probably got it from sbcl