Do not overwrite the system's idea of what the tmpdir is.
/tmp might be a separate directory that turns out to be too small,
especially for our bitcode build.  I discovered that the reason why I
could never build bitcode in FreeBSD is that I have a smaller+faster
ramdisk on /tmp, and that the parallelism was hiding the real error
message (OS to terminal).

It is the user's responsibility to set shell environment variables
when /tmp is not suitable.  Overriding this in wscript would prevent
the user from doing that.

Also, I had one linker which didn't overstand one of the options when
I tried lotsa linkers.

Also, using a hardcoded path in /tmp is a security problem.