Translate some lexical locations as proper SSA
Basically use the LLVM values directly instead of an alloca with
a store, if it so happens that a lexical location is SSA (i.e. has
only one definer).

Does not work because code with tagbody and cells will generate
invalid LLVM-IR in that a defining instruction "will not dominate"
its users. Scarequotes because the logic never actually works that
way. LLVM thinks it does because of the one-per-function landing pad
that can end up at any tag. It's an obscure problem.

The definition of add-package-local-nickname demonstrates it.