Don't obsolete instances of barely redefined classes
We only have to obsolete instances of classes redefined so that the
actual instance structure, i.e. instance slots, change.

This is required to support DEFSTRUCT reload (and Iterate, see #619)
but also helps efficiency a bit in that we don't have to recompile
discriminating functions after some class redefinitions or reloads.