Synchronize BIND-VA-LIST outputs with lambda list
Fixes #638. That has some explanation, but to resummarize:
arguments.lsp works off the lambda list, which won't match
the actual outputs of the bind-va-list instruction in the
case that the bind-va-list's outputs have been setf'd after
the fact - in #638's case, due to one of the parameters being
closed over (cell-converted).