Separate literals and transient values
Literals are objects that the code refers to, transient values
are values that are used by the RUN-ALL to create the literals.
They are now separated into separate vectors. The literals are in
the literal vector and the transients are in a separate vector that
is garbage collected once it is finished with.