Fill in holes in source location information
The (ensure-origin origin 99999xxx) forms need to be removed
but for now they indicate source location information that is missing
if it comes through this code path with an origin of nil.
A source line like 999990 means a nil origin was discovered and
replaced with a bogus origin and that we need to identify
where the nil origin is coming from.