Clasp now provides source line info to slime!!!
With the latest slime (I'll push that next)
hit 'v' in slime and it will jump to the form leading
to the error.
I had a problem where llvm merged code paths that
set important source line information to zero.
This change: (llvm-sys:set-enable-tail-merge target-pass-config nil)
Disables the tail-merge (BranchFolding) machine optimization
and avoids the problem at the expense of a little more code being
generated.