Fix .method-args. unboundedness in satiated GFs
As explained in the comment, the call history can be out of sync
with the dtree. If the call history doesn't need a vaslist but the
dtree does, the lambda will have fixed arguments but the outcomes
will refer to a vaslist. This causes warnings and errors.

The particular manifestation I ran into was in only one function,
SUBSTITUTE-OUTPUT, and only during a parallel build.