Fix proxy source info
This is used for e.g. the C-c C-c swank-compile-string mechanism
in SLIME: We give compile-file a temp file, but want the debug
information to reflect a concrete file. The temp file has some
portion of the actual file at some offset.

This change adds line number information and also fixes the file
name, which was apparently broken. The file name is not controlled
by the source-pos-info directly, resulting in my confusion.