Rethink MAP optimization
As per trends elsewhere, gonna not open-code so as to avoid code
size blowup. The main things we want to optimize with MAP are
a) the result-type processing - that's not much different, and
b) avoiding the variadic case where we have to cons up a list of
   iterators and so on.
So for b we specialize the one-sequence case, which is most common
by a long shot according to my cursory search, and leave the rest
in the lurch, at least for now.
Using the sequence iterators shouldn't really be any slower than
repeatedly calling vref - maybe even faster if we spice up
make-sequence-iterator to return element-type-specialized
functions. TODO