Responding to change with the 'nm' utility
It looks like Apple did something to 'nm' or libraries
because now the symbols that it dumps are different and confusing clasp.
Clasp uses 'nm' to read symbol addresses in the libraries that it
loads at startup. For some reason we are getting very long symbols so I
switched from using fgets to getline.
It's slower now too - we will have to look into this.