Fix setf for hash-tables.
This was a tricky bug that hit hash-tables when using MPS.
When evaluating a (setf (gethash key hashtable) value)
If the lookup for the entry reference in the hashtable hit
a stale pointer in the key, the hashtable would be rehashed.
In the rehash it searches for the key and returnes a reference
to the entry in the OLD table, not the new rehashed one!!!!
So I fix this by looking up the reference for the key
at the end of the rehash.