Delete no-longer-function dynamic variable watch code
If we want to do this in the future, here's how I'd do it:
Add a flag to Symbol_O (rather than using an alist like this
commit deletes). Add a branch in set_threadLocalSymbolValue to
check it, and call a function or what the fuck ever. That function
should be called any time a special is bound or unbound.