Removal nonlocal effects in calculating maybe-entry landing pads
While debugging I ran into a case where code that never executed
influenced the compilation of the code that did - it had a BLOCK
that was nonlocally returned to. This is okay, but kind of weird
and I don't like it. So now we check just for entries in the
specific context. Cleaner anyway.