Fix subclassing of standard-class
The specified :locations were wrong after I deleted a bunch of
slots. I fixed them and added a sanity check.

We might want to reexamine the :location mechanism. For one thing
we have, like, two different functions doing it? For two there are
some ambiguous points, like if a location is higher than the
slot count, should we inflate the slot count artificially, or
signal an error?