Fix defcallback in compile-file-parallel #954
The problem was that clasp-cffi::%expand-callback-definition
was looking up the llvm types during the defcallback
macro expansion - which is too early for the parallel compiler.
Looking up the llvm types during macro expansion is a bad idea
because it happens in the main compilation thread and in the
parallel compiler the llvm Module's are constructed in
child threads. Each thread has its own LLVMContext and
you can't put types from one LLVMContext into a
Module with another LLVMContext.

So now I pass the type symbols all the way through to the HIR
instructions and convert them to llvm types in the child threads.