parent thread passes less info to child thread
Previously I was passing a pointer to a Process_O object
from the parent to the child - but that was bad because
with MPS moving GC the Process_O object passed could move.

So now I pass an 64-bit ID that the child uses to find
the Process_sp in the systems list of processes.