Fix single dispatch generic functions
Now we use FuncallableInstance_O and a few slots
in the Rack to store a call-history for single dispatch
generic functions that support calling C++ methods.
The call-history is used to dispatch to methods
for single dispatch generic functions.
This moves us closer to compiling discriminating functions
for single dispatch generic functions.