Use compiler macros instead of inline definitions for some stuff.
Compiler macros are a lot cheaper than inline definitions for the
compiler to handle. The idea is that if the transformation is simple
enough (i.e. the argument forms are only used once, and hence can be
expressed this way), the compiler macro is preferable to an inline
declaration, especially since bclasp can also use it.