Fix equalp on single-float with different type
Previously (equalp single bignum) and (equalp single double)
always returned true, as far as I can tell.

I think this bug was introduced in
2348844606c8c408b802cda155b9b32f1158a19d way back in 2015, but if
so, it's kind of crazy that we haven't noticed before. I guess
equalp isn't actually used that often, and less often on numbers.

Fixes #1110 which was indirectly caused by this.