Ran static analyzer
Then commented out a few lines of
clasp_gc_cando.cc so that it compiles.
There are still private fields that cause compile time problems
but they are a symptom of a deeper problem that there are C++
stl data types being used as fields in GC managed objects.
They must be eliminated to allow those classes to work with
snapshot save/load