Remove stack_top_hint
In the high level debugger this can be done with-capped-stack; the
only difference is that it will compute all the frames even past
the cap and only show the capped ones, unlike stack_top_hint, which
would actually stop computing the frames. But the efficiency
difference is pretty marginal.

In low level debugging we pretty much want to see all available
information anyway.