Generate GC table entries in correct order
Neither of these tables seem to be actually used at the moment,
and drmeister has indicated that they are old code that may be
useful again later. The still-extant obj_finalize function uses
the stamp as the index into the table.

Could be slightly more efficient at generation time by reusing the
sorted list and stuff.