Ensure ABORT and MUFFLE-WARNING never return
Per CLHS Arguments and Values for both, they never return. But a
dumb restart-bind can result in restarts that don't transfer
control and so do actually return some value.

There's a WSCL issue for this saying we should signal a control
error in this circumstance, which seems like a good idea.

Should fix #1179