New test framework
This is more like ansi-tests in that it has "expected" values and
displays the problem in more detail when a test fails.

I also deleted or altered several tests. Mostly stuff that
excessively relied on internals or didn't test all values.

Work in progress. I'm getting nasty looking compiler errors.