Rewrite representation selection
Now it can handle unboxed singles. Doing this with the old way of
a bunch of specific instructions was really messy. Having one
do-it-all instruction is easier, and probably okay since they're
only inserted after almost every other pass anyway.

There might be a slight issue if an argument is used directly as
an input to a thing accepting unboxed values, because there's
nothing to insert a cast since there's no outputting instruction.
In practice this doesn't happen because variables turn into
fixed-to-multiple.