Use safer and more specific unboxing
from_object_whatever does lots of coercions, which is too
complicated -we want either the right kind of float or an error.

On the flipside, inlining single float unboxing without a check
will give you garbage (though not segfaults or anything) if a non
single float is provided, which is no good either.

Calling a function to unbox and box is probably not worse than
things would be without unboxing by the compiler at all, so
I'm not too worried about the efficiency here.