Define some simple arithmetic identity reductions
I imagine it's pretty uncommon to write realpart calls where the
argument is known to be a real, and so on, so this is marginally
useful at best. Mostly I just wanted to make sure the transforms
can do these kinds of transformations for the sake of cl:float,
and to make sure they work for non float types.