Fix #1142 maybe
As partly described in that issue, the problem is that the
following sequence takes place during boot:

1) Symbols are allocated but not put in packages.
2) Packages are finished.
3) Symbols are finished, including putting them in the package.

Step 2 calls SHADOW on the package shadow list. Because the
symbols allocated are not yet in the package, this creates new
shadowing symbols. Then step 3 adds the old symbols, causing
duplication. I eliminated that part of step 2.

The "maybe" is because I don't entirely understand how this works:
it looks like despite adding a definition of LLVM-SYS:FUNCTION as
shadowing, this doesn't affect symbols_scraped_inc.h. But it does
seem to work, so I don't know what's going on there.