Fix phi representation selection wrt values
Using nil to mean both no values and "haven't seen a
representation yet" can cause conflict. In #1207, a phi node has
definitions with two obvious representations: (:object) and ().
If the () is seen first, with the old code, no change to the
phi's in-progress representation marker is made, so the subsequent
(:object) is used as the representation. This is incorrect, as it
may be that zero values are returned. The bug is nondeterministic
since there is no guarantee of whether (:object) or () is seen
first, and there is no problem if (:object) is seen first.

The fix introduces an out-of-band marker instead. Seems to work.
Not sure if this needs to be done for representation of variables
as well, since variables are always one-value really.

Fixes #1207.