Fix #1215
In values-collect-multi, if the variable form had no return values
memmove would be told to copy 0-1 t* from the values vector.
That's the huge number that is ~0, i guess, so a segfault resulted.
This bug seems to have been in place since at least the last time
i rewrote values-collect-multi, but didn't come up until
this kind of multiple-value-call was used by the constant folding.