Compile APPLY as MULTIPLE-VALUE-CALL
As described in the comment, this is useful because the compiler
can analyze mv-calls itself. For example, the optimization where
APPLY with a final NIL is turned into a call happens naturally as
part of the optimization of MV-CALL with a known number of values.

From quick tests I _think_ this is somewhere between slightly
faster and just as fast on non optimizable code, but I'm keeping
it in a branch for now just in case.