Inline APPLY
This (and the previous change) allows the compiler to MV-CALL
functions by allocating the arguments in a vector on the stack,
and passing this to the general entry point. This seems to be
something like 25%-50% faster depending on circumstances. In
particular, along with the vaslist transformation this can result
in no allocation being done at all, as in
(lambda (&rest r) ... (apply f r) ...) type code.