Declare _setjmp returns_twice to LLVM
I believe this is a longstanding bug that may have caused a
variety of hard to understand problems. The symptoms resemble
reading undefined data. My particular example was in
eclector.reader::read-rational, where the closed over
read-suppress boolean could end up as an integer (presumably a
pointer, but I'm not sure).