Guard against undumpable source forms for stepper
This could in rare situations result in COMPILE-FILE failing with
OPTIMIZE DEBUG but not without it. Specifically, if the source has
some undumpable literal object, but that object does not need to
be dumped normally (e.g. it is macroexpanded away), we need to be
just a little more careful.