Simplify logic in bundle
Finding the various directories in the bundle is done by
first looking for a src directory in the same directory as
the executable. This indicates that we are in a development
tree. If this is found then we navigate upwards to the
_SourceDir, which is the root of the tree.

If we are not in a development tree then check the environment
variable CLASP_HOME. If it is set then use that for _SourceDir.

If _SourceDir has been set by one of the previous two methods
then use the CLASP_DEV_yyy defines to locate the other
directories. If _SourceDir has not been set then use the
CLASP_INSTALL_yyy defines to locate all directories including
_SourceDir.

- Remove _InstallDir and _ContentsDir as they are not needed
  in updated algorithm.
- Rename _LispSourceDir to _SysDir to match its logical host
  name.
- Add logical host STARTUP for extension startup load scripts
  so they can be installed in a different location then the
  binaries.
- Add CLASP_DEV_yyy and CLASP_INSTALL_yyy C macros so that
  location of the various directories can be controlled by
  waf or koga.
- Add bin-path, lib-path and share-path slots to koga.