Move dynenv linkages to stack
When running ROOM, among other things, problems happen if an
object's pointers are dead references. This can happen if a dynenv
is allocated on the heap (as e.g. block dynenvs are) but points
into a parent dynenv on the stack that has since expired. While
the runtime does not really use this stale pointer, the GC still
expects it to be valid.

So instead, we store a list of dynenvs in stack allocated conses.
This takes the "outer" reference out of the block dynenvs, meaning
all their pointers are always valid. And it means we don't have to
do extra invalidation work, or clear memory - the conses are just
deallocated with the rest of their frames.