Only trip dynenv debug failure on stack survival
Dynenv conses can validly be "out of order" if they're in the same
stack frame. LLVM may decide to put our allocas on the stack in
some arbitrary order, not necessarily the order in which they
exist in the LLVM-IR. So instead what we want to look out for is
dynenvs that are from deallocated stack frames, i.e. that are on
the stack past the stack pointer. There's no easy way to get the
stack pointer from C++ (we could use asm), but the frame pointer
for the frame of the test function is a reasonable approximation.