Give up on "minimal" safety for the time being
and just check all types instead. This is because the minimal
checks are pretty nontrivial to work with once you start doing
more complicated type derivation through functions; you kind of
need to infer types backwards, i.e. given that you need the result
of a function to be some type, what types do the arguments need?
But that's difficult or impossible to work out in many cases.