Inline some simple T vector accesses
Unlike the old inlining based version, this time I was a bit more
careful to make sure we don't expand into more code. Basically
here we start with one lisp call to AREF, and replace that with
one C call to cc_checkBound followed by an array read/set, so not
so much the bloat we used to have. This looks about a third faster
in microoptimizing tests. Plus we can skip the bounds check in
unsafe code, and once I code it in it should be possible to skip
the bounds check if we prove it's unnecessary.