Introduce BOOLEAN rtype
Doing this with an eye towards removing the ugly magical transforms
concerning comparisons.

I like the rtype because it means no changes to the IR as far as
Cleavir optimizations are concerned, but allows us to express some
things a little more naturally - like no special conditional test
translations that all jump an i1 but in special ways. And, bonus,
maybe we'll be able to have code like
(let ((x (< a b))) (stuff) (if x ...)) avoid T/NIL comparisons.

The other bonus with removing the transformations will be in
reducing how critical if-if elimination is in meta-evaluate.
I mean, if-if elimination is good, but the less we need to rely on
complicated IR writing the better, IMO.