Fix stack leak with GO and make NLX more efficient.
This mathes what cvm does. We were forgetting to clean up the stack on
every GO. What's more, we _were_ cleaning up the stack on every entry
close, which is inefficient and unnecessary and precludes a more
efficient compilation strategy, which we implement here. We save a
push in the multiple value case and for the single value case we save
a push and pop by doing a jump for the normal control case, while
nothing changes for the "exceptional" control case.