Rebind landing pad while generating XEPs
The arguments parsing code uses call-or-invoke for the wrong-
number-of-arguments signal. Without this rebind, calling the
compiler from within the compiler (e.g. while compiling a
discriminating function that's part of the compiler) would result
in the invoke referring to the outer module's landing pad, which
is of course illegal and causes an LLVM verification failure.