VM: Do not call bytecode_vm recursively for special (un)binding.
We achieve this by allocating dynenvs on a separate stack. This is
only done for special binding dynenvs at the moment. We keep track of
the dynenv stack pointer and restore it on exit and entry, as well as
in C++ catch {} blocks.

We will want to do this for vm_exit and vm_entry as well next, since
that's where the biggest perforamcne improvement will come from.