Remove pessimizing moves
I don't actually totally understand how these are pessimizing;
the basic case for this warning seems to be about copy elision in
return values. My best guess is that it has something to do with
the fact that takeError is defined to take ownership anyway, so
I guess it is kind of a move operation already, and maybe returns
a prvalue?