Clean up reorder-ctor issues
The order of C++ initializer lists is not used - fields are
initialized in the order they are declared in the class. This
matters when an initializer relies on another initialization in the
same initializer list having been completed, but in general we
probably ought to keep the orders the same to avoid any confusion.