use source debug pathname in compile-file-parallel
This should fix the thing where SLIME complains about "wrong type
argument: characterp, nil" when you try to jump to source of a
frame that was compiled with C-c C-c. Using the truename means
that the frame source info will be the since-deleted tmpfile.
I have some changes for SLIME that should make it suck less even
when given a deleted file, also.