bytecode FASLs: inline module literals
Since it's not like they're gonna be coalesced anyway.
After some basic profiling I noticed it the compiler and loader
seemed to spend a good amount of time with arrays, and all the
(setf row-major-aref) increased the fasl size a bit. I think this
change shaved about a second off of SLIME startup (9s -> 8s).