Start on new performance tracing mechanism for CLOS
debug-fastgf is not suitable since it logs everything and has to
be paranoid - better for debugging the infrastructure as a whole.

I noticed a lot more dispatch missing than I expected, so it's
helpful to be able to say "record when this function AND ONLY this
function misses", so i don't need to worry about bootstrapping or
print-object or slime methods or etc.