bytecode compiler: optimize NLX
This should be the last major optimization pass the bytecode
compiler will ever need to do - as anything more involved should
be left up to Cleavir.

Some refinements are possible, though. First off, we don't need to
exit just because a local entry is closed over elsewhere - we just
need to entry_close. Second, in cases where a block is not used at
all, which is very common due to how DEFUN and FLET/LABELS work,
we could emit nothing, not even a save_sp.