Fix multiple-value-setq
The evalmacros definition does not properly account for subforms
when symbol macros are in play. This has been masked by the
compiler macro in inline.lisp. But we should just use the right
definition right off.

I'm not sure why I had to change the VALUES setf expander, or
rather how it was working before since the problem seems to be in
using SETF before it's defined as a macro, so it's compiled as a
function call.