bytecode: add&use DUP opcode
this makes setq a little less ridiculous to generate - we just use
a stack spot instead of a local variable. This will reduce the
number of local variables needed by some functions.

Also makes things a little easier on bytecode-to-bir. It was
screwing things up in that the temporary variable was conflated
with earlier bound variables, which caused problems when the
variable was closed over.