Allocate VM stack outside of ThreadLocalState
Having a large array inside the VirtualMachine (and thus the
ThreadLocalState) seems to crash the linker on Mac. It's also just
sort of wonky.

Probably ideally we'd have a more sophisticated setup here with
mmap and guard pages and growth and yada yada, but this works ok.
There is a definite suboptimality in allocating the entire max
stack size all at once, but it's not _too_ big so maybe it's ok.
It also means that the GC will walk the entire stack for pointers,
including the inactive part, but Boehm is probably not capable
of doing something more complex anyway. Or if it is it's wizardry.