Fixed a snapshot save/load bug on MacOS
I figure that some small compile were generating zero length
literal vectors and on MacOS you can't have zero length symbols.
I hacked it so that a single word (8-byte) literal vector symbol on MacOS is considered
a zero length literal vector.  This may be a horrible cludge.
It shouldn't have been a problem because the literal vector is also
not being relocated into executable memory - I didn't chase down why
not.  So a combination of an 8-byte literal vector that is not
relocated into executable memory and running on MacOS is being ignored
by the machinery that handles literal vector reconstitution from snapshots.