Optimize FUNCALL function
Normally it's not called much... except the bytecode compiler
is quite stupid so it will leave common code like
(FUNCALL #'(SETF AREF) ...) as an actual call to FUNCALL. So we
might as well optimize it. Of course, it might be good to soup up
the bytecode compiler a bit as well.

The main thing is using a vaslist instead of consing up a whole
stupid list.