cmpltv: limit moved initializations
Not very sure about this!
There was a problem with fcells and make-load-form order tests in
ANSI. The cause is that an initializer form included references to
(SETF FOO), which would be created by a sequence like
CONS -> RPLACA -> RPLACD -> FCELL-LOOKUP.
But the motion code for MLF would move the RPLACs to after the
FCELL-LOOKUP, screwing it up. This fix limits the moved
instructions to actual calls. Let's see if it works.
It would probably be good to come up with a more coherent
mechanism for dealing with this. The ordering is difficult.