Make Eclector use our native backquote
eclector.reader:quasiquote is fine (a bit underoptimized but who
cares), but it's not available really early. The particular
problem I hit is that if we save macro names that are expanded
(see subsequent commit) we end up referring to the ECLECTOR.READER
package during startup, before eclector is loaded, causing problems.